+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchanting/EnchantingFeature.java
@@ -4,12 +4,14 @@ import insane96mcp.iguanatweaksexpanded.IguanaTweaksExpanded;
 import insane96mcp.iguanatweaksexpanded.module.Modules;
 import insane96mcp.iguanatweaksexpanded.setup.SRRegistries;
 import insane96mcp.iguanatweaksexpanded.setup.registry.SimpleBlockWithItem;
+import insane96mcp.iguanatweaksexpanded.utils.LogHelper;
 import insane96mcp.iguanatweaksreborn.module.experience.anvils.Anvils;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.insanelib.base.config.LoadFeature;
+import insane96mcp.insanelib.data.lootmodifier.InjectLootTableModifier;
 import net.minecraft.ChatFormatting;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.screens.inventory.AnvilScreen;
@@ -23,12 +25,14 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.flag.FeatureFlags;
 import net.minecraft.world.inventory.MenuType;
+import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockBehaviour;
+import net.minecraftforge.common.data.GlobalLootModifierProvider;
 import net.minecraftforge.event.AnvilUpdateEvent;
 import net.minecraftforge.event.entity.player.ItemTooltipEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -51,6 +55,9 @@ public class EnchantingFeature extends Feature {
     @Label(name = "No enchanted smithing", description = "Enchanted items can no longer be upgraded (e.g. netherite)")
     public static Boolean noEnchantedSmithing = true;
 
+    public static final RegistryObject<Item> CLEANSED_LAPIS = SRRegistries.ITEMS.register("cleansed_lapis", () -> new Item(new Item.Properties()));
+    public static final RegistryObject<Item> ANCIENT_LAPIS = SRRegistries.ITEMS.register("ancient_lapis", () -> new Item(new Item.Properties().fireResistant()));
+
 	public static final RegistryObject<MenuType<SREnchantingTableMenu>> ENCHANTING_TABLE_MENU_TYPE = SRRegistries.MENU_TYPES.register("enchanting_table", () -> new MenuType<>(SREnchantingTableMenu::new, FeatureFlags.VANILLA_SET));
 
 	public EnchantingFeature(Module module, boolean enabledByDefault, boolean canBeDisabled) {
@@ -64,8 +71,14 @@ public class EnchantingFeature extends Feature {
 
     @SubscribeEvent
     public void onAnvilUpdate(AnvilUpdateEvent event) {
-        if (!this.isEnabled()
-                || !noEnchantmentMerge)
+        if (!this.isEnabled())
+            return;
+        preventMergingEnchantedItems(event);
+        cleansedLapis(event);
+    }
+
+    public void preventMergingEnchantedItems(AnvilUpdateEvent event) {
+        if (!noEnchantmentMerge)
             return;
 
         if (event.getRight().isEnchanted() || (event.getLeft().isEnchanted() && event.getRight().is(Items.ENCHANTED_BOOK)))
@@ -81,6 +94,24 @@ public class EnchantingFeature extends Feature {
         int rarityCost = Anvils.getRarityCost(enchantment);
         //return (float) (1 + rarityCost * lvl * (Math.pow(lvl, 0.25f)));
         return rarityCost * lvl + (lvl / 4f);
+    }
+
+    public void cleansedLapis(final AnvilUpdateEvent event) {
+        ItemStack left = event.getLeft();
+        if (!left.isEnchantable() || left.isEnchanted())
+            return;
+
+        ItemStack right = event.getRight().copy();
+        for (Enchantment enchantment : ForgeRegistries.ENCHANTMENTS.getValues()) {
+            LogHelper.info(enchantment.getDescriptionId() + ": " + enchantment.getMinCost(1));
+        }
+        if (!right.is(CLEANSED_LAPIS.get()))
+            return;
+        event.setCost(0);
+        event.setMaterialCost(1);
+        ItemStack result = left.copy();
+        left.getOrCreateTag().putBoolean(EnchantingFeature.INFUSED_ITEM, true);
+        event.setOutput(result);
     }
 
     @SubscribeEvent
@@ -115,5 +146,11 @@ public class EnchantingFeature extends Feature {
                 }
             }
         }
+    }
+
+    private static final String path = Modules.Ids.EXPERIENCE + "/Enchanting";
+    public static void addGlobalLoot(GlobalLootModifierProvider provider) {
+        provider.add(path + "blocks/lapis_ore", new InjectLootTableModifier(new ResourceLocation("minecraft:blocks/lapis_ore"), new ResourceLocation(IguanaTweaksExpanded.RESOURCE_PREFIX + "blocks/injection/cleansed_lapis")));
+        provider.add(path + "blocks/deepslate_lapis_ore", new InjectLootTableModifier(new ResourceLocation("minecraft:blocks/deepslate_lapis_ore"), new ResourceLocation(IguanaTweaksExpanded.RESOURCE_PREFIX + "blocks/injection/cleansed_lapis")));
     }
 }